Use reference counting to avoid over-pruning trie
There is an example of where the current pruner incorrectly removes a
necessary node hash, because it got duplicated. See the new test:
test_hexary_trie_avoid_over_pruning()

Added an incremental reference counter, which only prunes nodes when the
number of usages drops to zero. It only works on fresh databases. With
existing databases, it would still delete required nodes.

Also, skip an unnecessary database write, which now incorrectly
increments the reference counter, and causes a failure in the reference
counter.