Bugfix: stop deleting wrong key if given key gone
See the included test case, where a `del trie[b'']` was deleting the
trie[b'\x01'] key!

If you call HexaryTrie.delete(key) and key is missing, and the node
where the missing key would have been is an extension, then the
trie was deleting whatever child value was stored in that extension.