Make iter test much faster, run more examples
The deadline was occasionally hitting >600ms in CI, so it seemed worth a
doing a rewrite.

Also, hypothesis exposed a quirk: that you can't iterate to the b'' key
using NodeIterator.next(key)

Fixed that issue as well, by adding a key=None option that returns the
first possible key.