Add trie walk test of partial leaf traversal
Recently, there was a fix for incorrectly raising a TraversedPartialPath
when traversing into a leaf, even if the suffix of the leaf did not
match the traversal path. See:
https://github.com/ethereum/py-trie/pull/114

This bug was actually found during a hypothesis run during a trie walk.
Although there are new unit tests making sure the leaf traversal works
properly, it seems worth adding the example explicitly as a kind of
backup test that the scenario works properly during a trie walk &
backfill.