Bugfix no-op squash_changes on a pruning trie
The empty ref-count was making the squash_changes think it should prune
the root node, even if it was unchanged from the original trie.

This was caught when trying to walk over a trie that's pruning, and runs
a squash-changes, but sometimes doesn't change anything at all. It's
important to prune while walking because we want to have the trie
updates delete access to old nodes, to simulate losing access to old
trie nodes from peers that have moved on. This makes sure we don't
request trie nodes that don't exist anymore during backfill-only type
scenarios.