Add TraversedPartialPath.simulated_node
This is useful when traversing, to treat the trie "as if" there had been
a node at the place you traversed to.

For example, say you traverse to path (0xf, 0) into the root node, which
is an extension node. It has a key extension of (0xf, 0, 0xa, 9), so a
TraversedPartialPath is raised. The simulated node will be an extension
with key (0xa, 9), with the same child node.

Also adds the changelog entry, and updates the readme. (The readme had
the bug in plain sight all along! Can you spot it? Look for the change
in the fog prefixes that happened after using the new simulated_node
API).