Fix for walking trie with TraversedPartialPath
This fixes test_trie_walk_root_change_with_traverse(), which incorrectly
handled sub_segments during a TraversedPartialPath exception.

Tweak hypothesis to find the scenario: sometimes it helps to have nodes
stored to the database, instead of embedded. We can now force nodes to
be encoded, by increasing the minimum_value_length.

Now that we can reproduce the problem, fix the test. The chosen solution
is to use the new TraversedPartialPath.simulated node, and use the
sub_segments from that simulated node. This requires very little work
from the caller.