Use composite strategy in trie.tools.strategies
This is a cleanup for the backfill tests, using the shared strategy to
try to expand the surface area of the hypothesis tests.

Note:

- The trie_from_keys() builder with a minimum_value_length=0 would try
to "insert" at key b'' the value b'', which is of course a delete.  So
the smallest permissable size must be a byte length of 1.

- An empty trie is not handled well by these backfill tests, but
backfill is entirely uninteresting for an empty trie, so we can ignore
the case.